+++ packages/components/chip/src/Chip.stories.tsx
@@ -1,16 +1,20 @@
 import { Icon } from '@spark-ui/icon'
 import { AccountFill } from '@spark-ui/icons/dist/icons/AccountFill'
+import { ArrowHorizontalUp } from '@spark-ui/icons/dist/icons/ArrowHorizontalUp'
 import { CalendarOutline } from '@spark-ui/icons/dist/icons/CalendarOutline'
 import { Check } from '@spark-ui/icons/dist/icons/Check'
 import { EyeOutline } from '@spark-ui/icons/dist/icons/EyeOutline'
 import { MailOutline } from '@spark-ui/icons/dist/icons/MailOutline'
 import { Input as SparkInput, InputGroup } from '@spark-ui/input'
 import { Label } from '@spark-ui/label'
+import { Popover } from '@spark-ui/popover'
 import { VisuallyHidden } from '@spark-ui/visually-hidden'
 import { Meta, StoryFn } from '@storybook/react'
+import { cx } from 'class-variance-authority'
 import { ChangeEvent, ComponentProps, useState } from 'react'
 
 import { Chip } from '.'
+import { ChipLeadingIcon } from './ChipLeadingIcon'
 
 const meta: Meta<typeof Chip> = {
   title: 'Experimental/Chip',
@@ -19,7 +23,7 @@ const meta: Meta<typeof Chip> = {
 
 type ChipProps = ComponentProps<typeof Chip>
 
-const designs: ChipProps['design'][] = ['outlined', 'filled', 'dashed', 'tinted']
+const designs: ChipProps['design'][] = ['outlined', 'dashed', 'tinted']
 
 const intents: ChipProps['intent'][] = [
   'basic',
@@ -37,6 +41,38 @@ const intents: ChipProps['intent'][] = [
 export default meta
 
 export const Default: StoryFn = _args => <Chip>default chip</Chip>
+
+export const Kind: StoryFn = () => {
+  const [pressed, setPressed] = useState(true)
+
+  return (
+    <div className="flex flex-row gap-xl">
+      <Chip onClick={() => console.log('assist')}>
+        <ChipLeadingIcon>
+          <Icon label="calendar">
+            <CalendarOutline />
+          </Icon>
+        </ChipLeadingIcon>
+        <Chip.Content>Assist</Chip.Content>
+      </Chip>
+      <Chip pressed={pressed} onClick={(_, { pressed }) => setPressed(!pressed)}>
+        {pressed && (
+          <ChipLeadingIcon>
+            <Icon label="check">
+              <Check />
+            </Icon>
+          </ChipLeadingIcon>
+        )}
+        <Chip.Content>Filter</Chip.Content>
+      </Chip>
+      <Chip>
+        <Chip.Content>Input</Chip.Content>
+        <Chip.ClearButton label="clear" />
+      </Chip>
+      <Chip onClick={() => console.log('suggestion')}>Suggestion</Chip>
+    </div>
+  )
+}
 
 const singleSet = Object.entries({
   '🥝': ['fruit'],
@@ -64,7 +100,6 @@ export const SingleSelectionFilter: StoryFn = () => {
               key={filter}
               value={filter}
               pressed={isActive}
-              design={isActive ? 'filled' : undefined}
               onClick={(_event, { pressed, value }) => {
                 setActive(pressed ? undefined : (value as string))
               }}
@@ -126,7 +161,6 @@ export const UnionFilter: StoryFn = () => {
               key={filter}
               value={filter}
               pressed={isActive}
-              design={isActive ? 'tinted' : 'dashed'}
               onClick={(_event, { pressed, value }) => {
                 setActiveFilters(
                   pressed
@@ -201,7 +235,6 @@ export const IntersectionFilter: StoryFn = () => {
               key={filter}
               value={filter}
               pressed={isActive}
-              design={isActive ? 'filled' : 'dashed'}
               onClick={(_event, { pressed, value }) => {
                 setActiveFilters(
                   pressed
@@ -299,7 +332,7 @@ export const MaxWidth: StoryFn = () => {
         <Chip>{content}</Chip>
         <Chip onClear={() => console.log('clear')}>
           <Chip.LeadingIcon aria-label="label">
-            <Icon>
+            <Icon label="check">
               <Check />
             </Icon>
           </Chip.LeadingIcon>
@@ -311,6 +344,15 @@ export const MaxWidth: StoryFn = () => {
         </Chip>
         <Chip onClear={() => console.log('clear')}>
           <Chip.Content>{content}</Chip.Content>
+          <Chip.ClearButton label="clear" />
+        </Chip>
+        <Chip onClear={() => console.log('clear')}>
+          <Chip.Content>{content}</Chip.Content>
+          <Chip.TrailingIcon aria-label="label">
+            <Icon label="check">
+              <Check />
+            </Icon>
+          </Chip.TrailingIcon>
           <Chip.ClearButton label="clear" />
         </Chip>
       </div>
@@ -348,6 +390,27 @@ export const ActionIntent: StoryFn = _args => (
             key={`${design}-${intent}`}
             intent={intent}
             onClick={() => console.log(`click ${design} ${intent}`)}
+            onClear={() => console.log('clear')}
+          >
+            <Chip.Content>{intent}</Chip.Content>
+            <Chip.ClearButton label="clear" />
+          </Chip>
+        ))}
+      </div>
+    ))}
+  </div>
+)
+
+export const SelectionIntent: StoryFn = _args => (
+  <div className="flex flex-col flex-wrap gap-md">
+    {designs.map(design => (
+      <div key={design} className="flex flex-wrap gap-md">
+        {intents.map(intent => (
+          <Chip
+            defaultPressed={true}
+            design={design}
+            key={`${design}-${intent}`}
+            intent={intent}
             onClear={() => console.log('clear')}
           >
             <Chip.Content>{intent}</Chip.Content>
@@ -478,5 +541,76 @@ export const Suggestion: StoryFn = () => {
         </Chip>
       </Component>
     </div>
+  )
+}
+
+const animals = ['Cat', 'Dog', 'Bird', 'Snake']
+export const Selection: StoryFn = () => {
+  const [isPressed, setIsPressed] = useState(false)
+  const [isOpened, setIsOpened] = useState(false)
+  const [value, setValue] = useState<string[]>([])
+  const hasValues = Boolean(value.length)
+  const content = value.length === animals.length ? 'All' : value.join(', ')
+
+  return (
+    <Popover open={isOpened}>
+      <Popover.Trigger asChild>
+        <Chip
+          pressed={isPressed || hasValues}
+          onClick={(_, { pressed }) => {
+            setIsOpened(!isOpened)
+            setIsPressed(!pressed)
+          }}
+          onClear={() => {
+            setValue([])
+            setIsPressed(false)
+          }}
+        >
+          <Chip.Content>{hasValues ? content : 'Animals'}</Chip.Content>
+          {value.length ? <Chip.ClearButton label="clear" /> : null}
+          <Chip.TrailingIcon>
+            <Icon
+              size="sm"
+              label={isOpened ? 'opened' : 'closed'}
+              className={cx('duration-150 ease-out', isOpened ? 'rotate-0' : '-rotate-180')}
+            >
+              <ArrowHorizontalUp />
+            </Icon>
+          </Chip.TrailingIcon>
+        </Chip>
+      </Popover.Trigger>
+      <Popover.Portal>
+        <Popover.Content align="start" asChild onInteractOutside={() => setIsOpened(false)}>
+          <ul className="overflow-hidden p-none">
+            {animals.map(arrayValue => {
+              const isValueIncluded = value.includes(arrayValue)
+
+              return (
+                <button
+                  role="listitem"
+                  key={arrayValue}
+                  className={cx(
+                    'flex w-full flex-row items-center justify-start gap-sm px-lg py-md',
+                    isValueIncluded
+                      ? 'bg-basic text-on-basic'
+                      : 'bg-basic-container text-on-basic-container',
+                    isValueIncluded ? 'hover:bg-basic-hovered' : 'hover:bg-basic-container-hovered',
+                    isValueIncluded ? 'focus-visible:bg-basic' : 'focus-visible:bg-basic-container'
+                  )}
+                  onClick={() =>
+                    setValue(
+                      isValueIncluded ? value.filter(v => v !== arrayValue) : [...value, arrayValue]
+                    )
+                  }
+                >
+                  <Icon label="check">{value.includes(arrayValue) ? <Check /> : <span />}</Icon>
+                  {arrayValue}
+                </button>
+              )
+            })}
+          </ul>
+        </Popover.Content>
+      </Popover.Portal>
+    </Popover>
   )
 }