+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.18](https://github.com/adevinta/spark/compare/@spark-ui/switch@2.0.17...@spark-ui/switch@2.0.18) (2023-09-13)
+
+**Note:** Version bump only for package @spark-ui/switch
+
 ## [2.0.17](https://github.com/adevinta/spark/compare/@spark-ui/switch@2.0.16...@spark-ui/switch@2.0.17) (2023-09-11)
 
 ### Bug Fixes